refactor(tests): nest streaming service tests under services/streaming
