+++ tests/unit/renderer/infrastructure/services/streaming/stream-track.monitor.test.ts
