+++ tests/unit/renderer/infrastructure/services/streaming/streaming.contract.test.js
